experiencein_back/requirements.txt
dj-database-url==0.5.0
Django==2.2
django-cors-headers==3.11.0
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
